Draft 1
Initiate the javascript, vue app, and adjust the location of index to do a test run. Verified the code userDate and userAge to make sure the random data is working but there will not be formatting errors. Everything else should be all set including profile 1 and 2 for test run. Profile page displays perfectly.